requirements.txt
beautifulsoup4==4.12.2
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
pytz==2023.3